switch scraper process to be on a different docker container
